fix(core): remove Node-only code to restore browser compatibility [closes #281]
Manifesto principle 1 requires @agentskit/core to work in any environment
(Node, Deno, edge, browser). The previous implementation imported
node:fs/promises at module load via config.ts, which crashed every
browser consumer at runtime with 'Module fs/promises has been
externalized for browser compatibility.'

This PR relocates the two Node-only helpers:

  createFileMemory  (packages/core/src/memory.ts)
    -> fileChatMemory  (packages/memory/src/file-chat.ts)

  loadConfig + AgentsKitConfig + LoadConfigOptions  (packages/core/src/config.ts)
    -> @agentskit/cli (packages/cli/src/config.ts)

What stays in @agentskit/core (all universal):
  createInMemoryMemory     works in every environment
  createLocalStorageMemory already guards on typeof localStorage

Consumer migrations:
- Re-exports of createFileMemory from @agentskit/react and @agentskit/ink
  are dropped (they were always Node-only; browser consumers got a
  broken export)
- apps/example-ink switched to fileChatMemory from @agentskit/memory
- @agentskit/cli internals now import loadConfig from its own ./config

Tests relocated:
- packages/core/tests/config.test.ts  -> packages/cli/tests/config.test.ts
- packages/core/tests/memory.test.ts createFileMemory block split out
  -> packages/memory/tests/file-chat.test.ts

Changeset declares:
  @agentskit/core          major (breaking — exports removed)
  @agentskit/memory        minor (new fileChatMemory export)
  @agentskit/cli           minor (new loadConfig public export)
  @agentskit/react         minor (dropped re-export)
  @agentskit/ink           minor (dropped re-export)

Verified:
- packages/core/dist/*.js contains zero fs/path imports
- apps/example-react mounts in a real browser (was empty root before)
- pnpm build succeeds for every package
- Ink test failure is pre-existing (ink@7 incompat fixed in #265),
  not caused by this change

Unblocks the 5 browser-side E2E tests currently skipped in #282.

Closes #281
