+++ apps/example-ink/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@agentskit/core": "workspace:*",
     "@agentskit/ink": "workspace:*",
+    "@agentskit/memory": "workspace:*",
     "ink": "^6.8.0",
     "react": "^19.2.5"
   },
