+++ packages/cli/src/commands.ts
@@ -2,8 +2,8 @@ import React from 'react'
 import { render } from 'ink'
 import { Command } from 'commander'
 import path from 'node:path'
-import { loadConfig } from '@agentskit/core'
-import type { AgentsKitConfig } from '@agentskit/core'
+import { loadConfig } from './config'
+import type { AgentsKitConfig } from './config'
 import { ChatApp, renderChatHeader } from './chat'
 import { writeStarterProject } from './init'
 import { runAgent } from './run'
