+++ packages/cli/tests/config.test.ts
